Added comments and documentation on how to create custom nodes
